pyproject.toml
[project]
name = "testcar"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "drf-yasg (>=1.21.9,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
